buildinfo: remove '/etc/build' and adapt sysinfo
All the information contained in this file can now be obtained from either
the system properties (using getprop) or from '/vendor/etc/repo-manifest.xml'

The sysinfo script has been adapted to these changes. Also, it will only show
the SHA1 of the repositories that belong to the 'sysinfo' group, instead of the
ones that have 'github-digi' or 'mtk' as remote.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>