device/digi: implement FAT boot and recovery sparse images
In previous versions of DEA, we were doing a hackish truncate of the
VFAT boot and recovery images to save space due to U-Boot not being able
to program sparse images. Now that U-Boot supports sparse images, just
use it for boot and recovery, just as we do for system and vendor images.

Also disable alignment (mkfs.vfat -a) when creating the FAT image. This
allows to have *one* reserved boot sector in the image, which is a
requirement for the 'fat16copy' script used afterwards.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>