dbts.mk: use TARGET_KERNEL_ARCH to build DTBs
TARGET_ARCH is the general variable used by Android for the configured
architecture. As we may need to build a mixed system with a 64-bit kernel
and DTBs but with 32-bit userspace, use a different kernel-specific
variable.

Also, use KERNEL_BIN as the dependence of the DTBs rule. Using
KERNEL_CONFIG is a bit brittle, as it can collide with the rule that
builds the kernel image.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>