ccimx8xsbcpro: delete dead code
The 'egl.cfg' files are not needed anymore since probably Android
Lollipop, so do not install those files.

Same for the 'init.imx8qm.rc' file, which is for an MX8 version we are
not using.

Also remove the commented lines that copy some tools scripts provided by
NXP to be used with fastboot or uuu. We are not using them at the
moment.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>