ccimx8xsbcpro: clean early initialization config files
NXP implements an init script 'init.insmod.sh' that allows to load
modules and set properties on target's boot. That script uses different
config files depending on the init's process stage when it's executed.

In our platform the only one executed from init is the 'early-init'
stage that uses 'early.init.cfg' config file. So remove the other unused
config files. Also from early-init config file remove the loading of the
wifi module, as we control it directly from Android settings.

Leave the file nevertheless, with some information about how to use it,
in case we need if for other modules or properties.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>