bootscript: remove 'swiotlb' kernel parameter
According to the NXP's Android user guide, the default value for this
parameter in the kernel is 64MB. They changed it to 96MB to workaround a
swiotlb overflow issue with a Wi-Fi driver. As we are not using the same
Wi-Fi driver that NXP, leave the default kernel value by removing the
parameter.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>