ccimx8xsbcpro/bootscript.txt: add new U-Boot setting to disable interfaces on sleep
The new SCU and Kernel have the ability to disable the 3.3V regulator when the device goes
to sleep. This causes the Wi-Fi/BT interfaces to fail to initialize after the device wakes up.
This commit adds a new U-Boot property that determines whether the Wi-Fi and Bluetooth
interfaces should be disabled before sleep and re-enabled on wake up to avoid initialization
errors.

https://jira.digi.com/browse/ADK4A-1767

Signed-off-by: David Escalona <david.escalona@digi.com>