build: add support for local configurations
This allows to do some build configurations using a 'localconf.mk'
makefile that is installed in the root build folder.

The localconf.mk customizations are put available to the rest of the
makefiles via the '01_localconf.mk' makefile. This file needs to keep
the lowest ordering number in the name, so it's included in the general
android build system *before* the rest of the makefiles (u-boot, kernel,
etc)

Mainly used for Trustfence configuration, so it is already filled with
TF defaults.

https://jira.digi.com/browse/ADK4A-1773

Signed-off-by: Javier Viguera <javier.viguera@digi.com>