device/digi: use Digi specific imx-make.sh shell script
We keep the script with the minimal changes from the one in 'device/nxp'
for easier maintenance.

This commit only changes the calculation of the path to the NXP and Digi
devices directory. We need both of them, because some makefiles are used
from Digi's folder while others are used from the NXP folder.

https://jira.digi.com/browse/ADK4A-1794

Signed-off-by: Javier Viguera <javier.viguera@digi.com>