ccimx8: init.rc: watchdogd: remove watchdog daemon initialization
Since the Watchdog API makes use of the watchdog character device, remove the
watchdogd daemon initialization so that the Digi API can access it. The watchdog
Digi service will start the system watchdog and periodically kick it until an
application registers the service and takes control over it.

Related to commits:
   * 6d867a19de13375acd9cb5957e781a1201a1d1cf
   * 5d968cf76ce040400c51fc0febcc29d01b26948c

https://jira.digi.com/browse/ADK4A-1795

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>