ccimx8mmdvk/config.xml: reduce the power wake lock on connection loss to 1 second
By default Android acquires a power lock of 60 seconds everytime the main connection is
lost in order to give it time to recover. This wake lock was preventing the device to
fully go to sleep when Wi-Fi or ethernet interface was shut-down before sleeping. This
commit reduces the wake lock to 1 second, allowing the device to correctly sleep after
interfaces are disabled.

Related to commit 7d38059dc9851d21cbd45806a515dd5cedaed096

https://jira.digi.com/browse/ADK4A-1768
https://jira.digi.com/browse/ADK4A-1830

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>