common: DigiServicesApp: update firmware targets file for the new API
Remove the system target and rename the manifest target to OTA JSON configuration file.
The entry point for the firmware udpate now is the JSON configuration file, there is
no way to update the system like it was done previously using the OTA zip file directly
or fragment it. The JSON configuration file describes the OTA location and the update
mechanism.

https://onedigi.atlassian.net/browse/ADK4A-1849

Signed-off-by: David Escalona <david.escalona@digi.com>