deploy script: allow to use 'dtbo.img' and 'vbmeta.img' image names
This commit adds to the uuu install script the possibility of using images
without a suffix with the processor.

Mainly to be used with the uncompressed zip file generated after singing images
with a custom key. These images does not have the suffix in the name.

The script first looks for the images with the prefix and, if not found, it
tries with 'dtbo.img' and/or 'vbmeta.img'.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>