ccimx6sbc: config.xml: reduce the power wake lock on connection loss to 1 second
By default Android acquires a power lock of 60 seconds everytime the main connection is
lost in order to give it time to recover. This wake lock was preventing the device to
fully go to sleep when Wi-Fi or ethernet interface was shut-down before sleeping. This
commit reduces the wake lock to 1 second, allowing the device to correctly sleep after
interfaces are disabled.

https://onedigi.atlassian.net/browse/ADK4A-1900

Signed-off-by: David Escalona <david.escalona@digi.com>