deploy script: add a new mode to program closed devices
The new mode, 'tf-deploy', is similar to 'first-deploy', but it needs an
additional bootloader image to be loaded via USB.

This image to load should not include trusty, since if the device is closed,
trusty tries to access the RPMB secure storage and fails.

Both bootloader images (to load and to program) must be properly signed.

https://onedigi.atlassian.net/browse/ADK4A-1896

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>